Update remaining accessors to `const`
Update the remaining accessors, namely those that return(ed)
mutable references, to const, by both marking the member function
as `const`, and changing the return type to `const &`.

The compilation error received without the modified return type
is of the form:

> binding reference of type ‘foo&’ to ‘const foo’ discards qualifiers

My understanding of the issue here is that returning a mutable
reference would allow (possible) modification of the callee
instance through that reference, thus bypassing the `const`
assertion of the member function.